Standardize quantity formatting to whole numbers across WGIMS
- Cast all quantity fields from float to integer in 9 models
  (Item, Delivery, DeliveryItem, DeliverySubsidy, DeliverySubsidyItem,
   RequisitionItem, RequisitionDispatchItem, StockCardEntry, StockTransferItem)
- Enforce integer validation for all quantity inputs in controllers
  (min:1 for new quantities, min:0 where full reversal is allowed)
- Display quantities as whole numbers (no decimals) across items,
  inventory balance, subsidies, requisitions/RIS, augmentations,
  partial deliveries, dispatching, stock transfers, stock cards,
  dashboards, reports, print views, modals and summary totals
- Keep 2-decimal precision for monetary values: unit cost,
  ENGAS unit cost, total cost, ENGAS total cost
- Update form inputs to step=1 for quantities, step=0.01 retained for costs
- Update JavaScript formatting (toLocaleString maximumFractionDigits: 0)
  for quantity displays; cost formatting untouched
- Excel exports: quantity columns use #,##0 format, cost columns keep #,##0.00
- Calculations unchanged: Total Cost = Quantity x Unit Cost and
  ENGAS Total Cost = Quantity x ENGAS Unit Cost use full monetary precision
- Update affected feature tests to expect whole-number rendering

+++ app/Http/Controllers/DashboardController.php
@@ -50,7 +50,7 @@ class DashboardController extends Controller
                     $accountBalances[] = [
                         'account_code' => $cat['account_code'],
                         'label'        => $cat['label'],
-                        'total_qty'    => (float) $row->total_qty,
+                        'total_qty'    => (int) $row->total_qty,
                         'total_value'  => (float) $row->total_value,
                     ];
                 }
@@ -116,7 +116,7 @@ class DashboardController extends Controller
             $accountBalances[] = [
                 'account_code' => $cat['account_code'],
                 'label'        => $cat['label'],
-                'total_qty'    => (float) $row->total_qty,
+                'total_qty'    => (int) $row->total_qty,
                 'total_value'  => (float) $row->total_value,
             ];
         }
